Extract common functionality to useRecentSearches composable
